requirements.txt
Pillow==10.1.0
numpy==1.26.2
cairosvg==2.7.1
Jinja2==3.1.2
